feat: Create necessary entities for business model
This commit introduces the creation of essential entities required for the business model implementation. The new entities have been defined and annotated according to the project's business logic. These entities are fundamental to representing various aspects of the online store's operations.

By adding these entities, we are laying the foundation for the business logic and data handling of the online store project. These entities encapsulate the core entities and relationships needed to build the functionality defined in issue #SCRUM-57.

Refs: SCRUM-57

+++ src/main/resources/db/migration/V1.1__create_product_table.sql
@@ -30,8 +30,7 @@ create table products
     country_producer varchar(255),
     exist            boolean       not null default true,
     priority_score   int           not null default 0,
-    rating           double (2, 1
-) ,
+    rating           double (2, 1) ,
     review_count int ,
     discount_percent int not null default 0 ,
     create_date datetime not null ,
